JustDate sql store milliseconds
git-svn-id: http://192.168.0.243:4435/svn/bee/trunk@591 aaee961a-53e7-2249-bd01-b2a8f0c4ce08

+++ src/com/butent/bee/shared/time/JustDate.java
@@ -222,7 +222,7 @@ public class JustDate extends AbstractDate implements Comparable<JustDate> {
   }
 
   public long getTime() {
-    return (long) TimeUtils.MILLIS_PER_DAY * days;
+    return TimeUtils.MILLIS_PER_DAY * days;
   }
 
   @Override
